requirements.txt
requests>=2.25.1
langchain_openai 
langchain_core
python-dotenv==1.0.1
pydantic_ai
pydantic-ai[logfire]
scrapegraphai
langchain-google-vertexai
beautifulsoup4>=4.11.0
lxml>=4.9.0
html5lib>=1.1
htmlmin
sendgrid
tavily-python
crawl4ai
psutil
FlagEmbedding
supabase
crewai
crewai_tools
openai
pygame